+++ src/crawlers/statistics.js
@@ -253,6 +253,15 @@ class Statistics {
 
         if (!savedState) return;
 
+        // We saw a run where the requestRetryHistogram was not iterable and crashed
+        // the actor. Adding some logging to monitor this problem in the future.
+        if (!Array.isArray(savedState.requestRetryHistogram)) {
+            this.log.warning('Received invalid state from Key-value store.', {
+                persistStateKey: this.persistStateKey,
+                state: savedState,
+            });
+        }
+
         this.log.debug('Recreating state from KeyValueStore', { persistStateKey: this.persistStateKey });
 
         this.requestRetryHistogram.push(...savedState.requestRetryHistogram);